# docs/openapi/approval_pack.json
{
  "info": {
    "title": "Pacha Claim Core",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/claims/{claim_id}/approval-pack/artifacts/{event_id}": {
      "get": {
        "operationId": "read_artifact_claims__claim_id__approval_pack_artifacts__event_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "claim_id",
            "required": true,
            "schema": {
              "title": "Claim Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor",
            "required": true,
            "schema": {
              "title": "X-Actor",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Artifact"
      }
    },
    "/claims/{claim_id}/approval-pack/generate": {
      "post": {
        "operationId": "generate_claims__claim_id__approval_pack_generate_post",
        "parameters": [
          {
            "in": "path",
            "name": "claim_id",
            "required": true,
            "schema": {
              "title": "Claim Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor",
            "required": true,
            "schema": {
              "title": "X-Actor",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate"
      }
    },
    "/claims/{claim_id}/approval-pack/manifest/{item_id}/sources": {
      "put": {
        "operationId": "select_sources_claims__claim_id__approval_pack_manifest__item_id__sources_put",
        "parameters": [
          {
            "in": "path",
            "name": "claim_id",
            "required": true,
            "schema": {
              "title": "Claim Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor",
            "required": true,
            "schema": {
              "title": "X-Actor",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceSelection"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Select Sources Claims  Claim Id  Approval Pack Manifest  Item Id  Sources Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Select Sources"
      }
    },
    "/claims/{claim_id}/approval-pack/manifest/{item_id}/upload": {
      "post": {
        "operationId": "upload_item_claims__claim_id__approval_pack_manifest__item_id__upload_post",
        "parameters": [
          {
            "in": "path",
            "name": "claim_id",
            "required": true,
            "schema": {
              "title": "Claim Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor",
            "required": true,
            "schema": {
              "title": "X-Actor",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_item_claims__claim_id__approval_pack_manifest__item_id__upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Upload Item Claims  Claim Id  Approval Pack Manifest  Item Id  Upload Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Item"
      }
    },
    "/claims/{claim_id}/approval-pack/note-drafts": {
      "get": {
        "operationId": "read_note_drafts_claims__claim_id__approval_pack_note_drafts_get",
        "parameters": [
          {
            "in": "path",
            "name": "claim_id",
            "required": true,
            "schema": {
              "title": "Claim Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor",
            "required": true,
            "schema": {
              "title": "X-Actor",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Read Note Drafts Claims  Claim Id  Approval Pack Note Drafts Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Note Drafts"
      }
    },
    "/claims/{claim_id}/approval-pack/readiness": {
      "get": {
        "operationId": "read_readiness_claims__claim_id__approval_pack_readiness_get",
        "parameters": [
          {
            "in": "path",
            "name": "claim_id",
            "required": true,
            "schema": {
              "title": "Claim Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor",
            "required": true,
            "schema": {
              "title": "X-Actor",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Read Readiness Claims  Claim Id  Approval Pack Readiness Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Readiness"
      }
    },
    "/claims/{claim_id}/approval-pack/versions": {
      "get": {
        "operationId": "read_versions_claims__claim_id__approval_pack_versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "claim_id",
            "required": true,
            "schema": {
              "title": "Claim Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor",
            "required": true,
            "schema": {
              "title": "X-Actor",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Read Versions Claims  Claim Id  Approval Pack Versions Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Versions"
      }
    },
    "/reviews/{review_id}/approval-note": {
      "get": {
        "operationId": "read_workspace_reviews__review_id__approval_note_get",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "title": "Review Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor",
            "required": true,
            "schema": {
              "title": "X-Actor",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Read Workspace Reviews  Review Id  Approval Note Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Workspace"
      }
    },
    "/reviews/{review_id}/approval-note/draft": {
      "put": {
        "operationId": "autosave_draft_reviews__review_id__approval_note_draft_put",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "title": "Review Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor",
            "required": true,
            "schema": {
              "title": "X-Actor",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutosaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Autosave Draft Reviews  Review Id  Approval Note Draft Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Autosave Draft"
      }
    }
  }
}